requirements.txt
timm
torch
fastapi
Pillow
uvicorn
python_multipart